New guide added (to match new ChillerCase_v2.amf)
New CHILLER 2.0 setup+assembly guide added, which matches the new case+lid file (ChillerCase_v2.amf). Contains instructions for assembling chiller with new case, which has rounded contours, an aperture for charging the power bank, room for an on/off switch, and accommodates the LED strip in the long direction.